refactor: non-facade method tests extracted
